+++ R/scoreci.R
@@ -249,9 +249,9 @@ scoreci <- function(
 	p2hat <- x2/n2
 	
 	#wrapper function for scoretheta
-	myfun <- function(theta, randswitch = tdas, ccswitch = cc) {
+	myfun <- function(theta, randswitch = tdas, ccswitch = cc, stratswitch = stratified) {
 	  scoretheta(theta = theta, x1 = x1, x2 = x2, n1 = n1, n2 = n2, bcf = bcf,
-	             contrast = contrast, distrib = distrib, stratified = stratified,
+	             contrast = contrast, distrib = distrib, stratified = stratswitch,
 	             wt = wt, weighting = weighting, tdas = randswitch, skew = skew,
 	             cc = ccswitch)$score  
 	}
@@ -306,7 +306,7 @@ scoreci <- function(
 	p2d.MLE <- at.MLE$p2d
 	wt.MLE <- at.MLE$wt
 	V.MLE <- at.MLE$V
-
+	
 	# if stratified=TRUE, options are available for assuming fixed effects (tdas=FALSE)
 	# or random effects (tdas=T). The IVS weights are different for each version, which 
 	# in turn can lead to a different point estimate, at which certain quantities are 
@@ -326,7 +326,7 @@ scoreci <- function(
 	  Q.FE <- at.FE$Q
 	  I2 <- max(0, 100 * (Q.FE - (nstrat - 1))/Q.FE)
 	  pval.het <- 1 - pchisq(Q.FE, nstrat - 1)
-	  # Qualitative interaction test to be added...
+	  # Qualitative interaction test is calculated further down
 	  
 	  # as per M&N p218 (little r), actually no longer needed for point estimate. 
 	  p1hat.w <- sum(wt.MLE * x1/n1)/sum(wt.MLE)  
@@ -365,6 +365,25 @@ scoreci <- function(
 	  myfun(theta) + qtnorm, contrast = contrast, distrib = distrib,
 	  precis = precis + 1, uplow = "up")
 
+	
+	#get estimate & CI for each stratum
+	if(stratified == TRUE) {
+	  at.MLE.unstrat <- scoretheta(theta = point, x1 = x1, x2 = x2, n1 = n1, n2 = n2, bcf = bcf,
+	                               contrast = contrast,
+	                               distrib = distrib, stratified = FALSE,
+	                               weighting = weighting, wt = wt, tdas = tdas, skew = skew,
+	                               cc = cc)
+	  point.FE.unstrat <- bisect(ftn = function(theta) 
+	    myfun(theta, randswitch = FALSE, ccswitch = 0, stratswitch = FALSE) - 0, contrast = contrast,
+	    distrib = distrib, precis = precis + 1, uplow = "low")
+	  lower.unstrat <- bisect(ftn = function(theta)
+	    myfun(theta, stratswitch = FALSE) - qtnorm, contrast = contrast, distrib = distrib,
+	    precis = precis + 1, uplow = "low")
+	  upper.unstrat <- bisect(ftn = function(theta)
+	    myfun(theta, stratswitch = FALSE) + qtnorm, contrast = contrast, distrib = distrib,
+	    precis = precis + 1, uplow = "up")
+	}
+	
 	# fix some extreme cases with zero counts
 	#if (contrast == "RR" && skew == FALSE) p2d.w[sum(x2) == 0] <- 0
 	
@@ -407,7 +426,10 @@ scoreci <- function(
 	} else if (contrast == "p") {
 	  delta0 <- 0.5
 	} else delta0 <- 1
-	if (is.null(delta)) delta <- delta0
+	if (is.null(delta)) {
+	  delta <- delta0
+	  qualdelta <- point
+	} else qualdelta <- delta
 	scorezero <- scoretheta(theta = delta0, x1 = x1, x2 = x2, n1 = n1, n2 = n2,
 	                        stratified = stratified,
 	                        wt = wt, weighting = weighting, tdas = tdas,
@@ -418,14 +440,30 @@ scoreci <- function(
 	                         wt = wt, weighting = weighting, tdas = tdas,
 	                         bcf = bcf, contrast = contrast, distrib = distrib,
 	                         skew = skew, cc = cc)
+	scorequal <- scoretheta(theta = qualdelta, x1 = x1, x2 = x2, n1 = n1, n2 = n2,
+	                         stratified = stratified,
+	                         wt = wt, weighting = weighting, tdas = tdas,
+	                         bcf = bcf, contrast = contrast, distrib = distrib,
+	                         skew = skew, cc = cc)
 	pval.left <- scoredelta$pval
 	pval.right <- 1 - pval.left
 	chisq.zero <- scorezero$score^2
 	pval2sided <- pchisq(chisq.zero, 1, lower.tail = FALSE)
 	if (tdas == TRUE) pval2sided <- pf(chisq.zero, 1, nstrat - 1, lower.tail = FALSE)
 	pval <- cbind(chisq = chisq.zero, pval2sided, delta = delta,
-	              scoredelta = scoredelta$score, pval.left, pval.right)
+	              scoredelta = scoredelta$score, pval.left, pval.right) #, sdot = at.MLE$Sdot)
 
+	#Add qualitative interaction test as per equation S4 of Laud 2017
+	Qc.i <- scorequal$Q.i
+	Qc.i.S3 <- scorequal$Q.i.S3
+#	Qc.i.S3 <- at.MLE$Q.i.S3
+	Qc <- min(sum(scorequal$Q.i.S3[scorequal$Stheta > 0]), sum(scorequal$Q.i.S3[scorequal$Stheta < 0]))
+#	Qc <- min(sum(scoredelta$Q.i[scoredelta$Stheta > 0]), sum(scoredelta$Q.i[scoredelta$Stheta < 0]))
+	Qcprob <- 0
+	for (h in 1:(nstrat - 1)) {
+	  Qcprob <- Qcprob + (1 - pchisq(Qc,h)) * dbinom(h, size = nstrat - 1, prob = 0.5)
+	}
+	
 	# Optional plot of the score function.
 	# Ideally this would be in a separate function, but it is unlikely to be used
 	# much in practice - only included for code development and validation purposes.
@@ -506,15 +544,16 @@ scoreci <- function(
 
 	outlist <- list(estimates = estimates, pval = pval) 
 	if (stratified == TRUE) {
-	  Qtest <- c(Q = Q.FE, tau2 = tau2.FE, pval.het = pval.het, I2 = I2)
+	  Qtest <- c(Q = Q.FE, pval.het = pval.het, I2 = I2, Qc = Qc, pval.qualhet = Qcprob) #tau2 = tau2.FE, 
 	  wtpct <- 100 * wt.MLE/sum(wt.MLE)
 	  wt1pct <- 100 * wt.FE/sum(wt.FE)
 	  outlist <- append(outlist,
 	    list(Qtest = Qtest, weighting = weighting, 
 	    stratdata = cbind(x1j = x1, n1j = n1, x2j = x2, n2j = n2,
-	                      p1hatj = p1hat, p2hatj = p2hat, Qj = Q.each,
-	                      wtpct.fixed = wt1pct, wtpct.rand = wtpct))) 
-#	    p1d=p1d.MLE,p2d=p2d.MLE,Stheta=Stheta.MLE,V.MLE)))
+	                      p1hatj = p1hat, p2hatj = p2hat, 
+	                      wtpct.fixed = wt1pct, wtpct.rand = wtpct, 
+	                      theta.j = point.FE.unstrat, lower.j = lower.unstrat, upper.j = upper.unstrat))) 
+#	  Qj = Q.each, Qc.j = Qc.i, Qc.j.S3 = Qc.i.S3,atmle = at.MLE$Stheta,  p1d=p1d.MLE,p2d=p2d.MLE,Stheta=Stheta.MLE,V.MLE, atdelta = scoredelta$Stheta, atmle = at.MLE$Stheta)))
 	}
 	outlist <- append(outlist, list(call = c(distrib = distrib,
 	                 contrast = contrast, level = level, skew = skew,
@@ -876,6 +915,7 @@ scoretheta <- function (
 		Sdot <- sum(wt * Stheta)/sum(wt)
 		if (weighting == "IVS") {
 		  Q.i <- wt * ((Stheta - Sdot)^2)  #This version for iterative weights?
+		  Q.i.S3 <- wt * ((Stheta)^2)  #This version for iterative weights?
 		} else Q.i <- ((Stheta - Sdot)^2)/V
 		Q <- sum(Q.i)
 #		Q[all(Stheta == Inf)] <- 0 #Attempt to get scoretheta(0) to work for contrast=="OR"
@@ -940,7 +980,7 @@ scoretheta <- function (
 	                p2d = p2d, mu3 = mu3, pval = pval)
 	if (stratified) {
 	  outlist <- append(outlist, list(Sdot = Sdot, Vdot = Vdot, tau2 = tau2,
-	             VS = VS, t2 = t2, Q.i = Q.i, Q = Q, wt = wt, p1ds = p1ds, p2ds = p2ds))
+	             VS = VS, t2 = t2, Q.i = Q.i, Q.i.S3 = Q.i.S3, Q = Q, wt = wt, p1ds = p1ds, p2ds = p2ds))
 	}
 	return(outlist)
 }